fix(gates): split dirty-pr-scan's base-branch parser into a sibling module
scripts/lib/dirty-pr-scan.mjs sits at 430 lines with this PR's new content,
which check-module-size.mjs's widened .mjs/scripts/ coverage (#3672/#3689,
live on current main) flags as a new file over the 400-line budget with no
allowlist row. Extract pullRequestBaseBranches and its private YAML-fragment
helpers (splitFlowList, stripQuotes, finishBranches) into
scripts/lib/workflow-base-branches.mjs: a self-contained unit -- parsing
test.yml's `on: pull_request: branches:` filter -- with no dependency on PR
classification. dirty-pr-scan.mjs re-exports pullRequestBaseBranches so
scan-dirty-prs.mjs and the existing test imports are unaffected. No behavior
change; the 43 existing tests pass unmodified.
